requirements.txt
streamlit==1.29.0
openai==1.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pandas==2.1.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-watch==4.2.0